+++ oak_functions/examples/metrics/module/src/lib.rs
@@ -19,7 +19,7 @@
 #[cfg_attr(not(test), no_mangle)]
 pub extern "C" fn main() {
     let request = oak_functions::read_request().expect("Couldn't read request body.");
-    if request == b"a" {
-        oak_functions::report_event("a").expect("Couldn't report event.");
-    }
+    let label =
+        std::str::from_utf8(request.as_ref()).expect("Request body is not a valid UTF-8 string.");
+    oak_functions::report_event(label).expect("Couldn't report event.");
 }
